Revise first and second year after meeting with CFM
CM had a few notes. First, they would like to continue offering the 101 class.
This means that the proposed CCE 101 class should be CCE 102.
But we agreed that it could be an optional course in both programs.
We also agreed that CCE 117 was not a necessary prequisite to CCE 270.
I have included a flowchart and revised text for both of these changes.